requirements.txt
numpy==1.26.4
pillow==10.4.0
scikit-image==0.24.0
torch==2.4.1+cu121
torchvision==0.19.1+cu121
